requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
sqlalchemy
psycopg2-binary
google-generativeai
langgraph
langchain
langchain-google-genai
redis
slowapi
python-multipart
structlog
black
flake8
mypy
pre-commit
